Time: 1 ms (86.03%), Space: 40.8 MB (68.8%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -3,7 +3,7 @@ class Solution {
         int left = 1, right = x;
         while(left <= right){
             int mid = left + (right - left) / 2;
-            if((long) mid * mid == (long) x){
+            if(mid * mid == x){
                 return mid;
             }
             else if((long)mid * mid < (long) x){
